Logic now controls Entity objects
The Logic system now implements an EntityFactory whose sole purpose is to produce an entity object when asked.  It will create and attach components to a newely created entity, so long as they are registered.  The LogicSystem will then house the entities as the exist through out the life-cycle of the engine..