Added ability to set up Game level Logic system
Games that implement the Alpha Engine should now implement their own version of the LogicSystem, and passing it as a template arguement to the InitializeAlpha static function.
This game level logic system should inherit from the base LogicSystem class, and all game level logic should be written here if c++ is needed.