Raw input translation and dispatch systems now setup
Each platform should inplement a window/input listener for raw input, and
pass that input to a translatort that will convert the raw code into an
engine code.

The engine code will then be dispatched, and picked up by the logic
system.  Where the logic system will run the engine code through a context
manager, and attempt to convert the engine code to a game specific input
mapping, e.g.: A maps to STRAFE_LEFT.

The next step is to allow logic system and game state implementations to
easily bing to game specific actions.  Also need to implement Linux
platform listener and translator.