Cleaned up demo project cmake to better support multiple projects
Moved Content up a directory so that it is shared by all demos. and it will now be copied once to the debug bin directory.
Added basic blank demo to show the minimum setup to get alpha up and running, and added a lights demo for the new lights feature once it is implemented.