Added support for light shaders
light shaders are separate from object shaders as they only emit light, and are not acted on my other lights in the shader.
Removed the LightComponent since it was overkill, and instead created a boolean in the base Scene Component the indicates if the object is a light emitter, or in other words, a light.  Since lights sometimes have meshes to draw on screen, this makes it easier to created say, a light bulb, that emits light in the scene.